building controller advice for error handling
