+++ src/main/java/rueppellii/backend2/tribes/building/web/BuildingController.java
@@ -2,7 +2,6 @@ package rueppellii.backend2.tribes.building.web;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import rueppellii.backend2.tribes.building.exception.UpgradeFailedException;
@@ -19,7 +18,6 @@ import rueppellii.backend2.tribes.progression.util.ProgressionDTO;
 import rueppellii.backend2.tribes.resource.exception.NoResourceException;
 import rueppellii.backend2.tribes.resource.service.ResourceService;
 import rueppellii.backend2.tribes.troop.exception.TroopNotFoundException;
-import rueppellii.backend2.tribes.user.util.ErrorResponse;
 
 import java.security.Principal;
 
@@ -65,62 +63,5 @@ public class BuildingController {
         resourceService.updateResources(kingdom);
         purchaseService.payForBuildingUpgrade(kingdom.getId(), building);
         progressionService.generateBuildingUpgradeModel(kingdom, id);
-    }
-
-
-    @ResponseBody
-    @ExceptionHandler(UsernameNotFoundException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    ErrorResponse userNotFoundHandler(UsernameNotFoundException ex) {
-        return new ErrorResponse(ex.getMessage());
-    }
-
-    @ResponseBody
-    @ExceptionHandler(KingdomNotFoundException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    ErrorResponse kingdomNotFoundHandler(KingdomNotFoundException ex) {
-        return new ErrorResponse(ex.getMessage());
-    }
-
-    @ResponseBody
-    @ExceptionHandler(TroopNotFoundException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    ErrorResponse troopNotFoundHandler(TroopNotFoundException ex) {
-        return new ErrorResponse(ex.getMessage());
-    }
-
-    @ResponseBody
-    @ExceptionHandler(BuildingNotFoundException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    ErrorResponse buildingNotFoundHandler(BuildingNotFoundException ex) {
-        return new ErrorResponse(ex.getMessage());
-    }
-
-    @ResponseBody
-    @ExceptionHandler(NoResourceException.class)
-    @ResponseStatus(HttpStatus.FORBIDDEN)
-    ErrorResponse NoResourceHandler(NoResourceException ex) {
-        return new ErrorResponse(ex.getMessage());
-    }
-
-    @ResponseBody
-    @ExceptionHandler(InvalidProgressionRequestException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    ErrorResponse InvalidProgressionHandler(InvalidProgressionRequestException ex) {
-        return new ErrorResponse(ex.getMessage());
-    }
-
-    @ResponseBody
-    @ExceptionHandler(IllegalArgumentException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    ErrorResponse InvalidProgressionEnumHandler(IllegalArgumentException ex) {
-        return new ErrorResponse(ex.getMessage());
-    }
-
-    @ResponseBody
-    @ExceptionHandler(UpgradeFailedException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    ErrorResponse InvalidUpgradeEnumHandler(UpgradeFailedException ex) {
-        return new ErrorResponse(ex.getMessage());
     }
 }
