+++ src/main/java/rueppellii/backend2/tribes/building/web/BuildingControllerAdvice.java
@@ -1,0 +1,75 @@
+package rueppellii.backend2.tribes.building.web;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import rueppellii.backend2.tribes.building.exception.BuildingNotFoundException;
+import rueppellii.backend2.tribes.building.exception.UpgradeFailedException;
+import rueppellii.backend2.tribes.kingdom.exception.KingdomNotFoundException;
+import rueppellii.backend2.tribes.progression.exception.InvalidProgressionRequestException;
+import rueppellii.backend2.tribes.resource.exception.NoResourceException;
+import rueppellii.backend2.tribes.troop.exception.TroopNotFoundException;
+import rueppellii.backend2.tribes.user.util.ErrorResponse;
+
+@ControllerAdvice(assignableTypes = BuildingController.class)
+public class BuildingControllerAdvice {
+
+    @ResponseBody
+    @ExceptionHandler(UsernameNotFoundException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    ErrorResponse userNotFoundHandler(UsernameNotFoundException ex) {
+        return new ErrorResponse(ex.getMessage());
+    }
+
+    @ResponseBody
+    @ExceptionHandler(KingdomNotFoundException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    ErrorResponse kingdomNotFoundHandler(KingdomNotFoundException ex) {
+        return new ErrorResponse(ex.getMessage());
+    }
+
+    @ResponseBody
+    @ExceptionHandler(TroopNotFoundException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    ErrorResponse troopNotFoundHandler(TroopNotFoundException ex) {
+        return new ErrorResponse(ex.getMessage());
+    }
+
+    @ResponseBody
+    @ExceptionHandler(BuildingNotFoundException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    ErrorResponse buildingNotFoundHandler(BuildingNotFoundException ex) {
+        return new ErrorResponse(ex.getMessage());
+    }
+
+    @ResponseBody
+    @ExceptionHandler(NoResourceException.class)
+    @ResponseStatus(HttpStatus.FORBIDDEN)
+    ErrorResponse NoResourceHandler(NoResourceException ex) {
+        return new ErrorResponse(ex.getMessage());
+    }
+
+    @ResponseBody
+    @ExceptionHandler(InvalidProgressionRequestException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    ErrorResponse InvalidProgressionHandler(InvalidProgressionRequestException ex) {
+        return new ErrorResponse(ex.getMessage());
+    }
+
+    @ResponseBody
+    @ExceptionHandler(IllegalArgumentException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    ErrorResponse InvalidProgressionEnumHandler(IllegalArgumentException ex) {
+        return new ErrorResponse(ex.getMessage());
+    }
+
+    @ResponseBody
+    @ExceptionHandler(UpgradeFailedException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    ErrorResponse InvalidUpgradeEnumHandler(UpgradeFailedException ex) {
+        return new ErrorResponse(ex.getMessage());
+    }
+}
